Fix signed/unsigned warning in VariableConverter. Somehow the deletion of core/TestSampleClient.h that was done in the Weffc++ branch didn't get applied to the trunk.

+++ src/nidas/core/VariableConverter.cc
@@ -446,7 +446,7 @@ void Polynomial::readCalFile(dsm_time_t t)
 {
     if (_calFile) {
         float d[MAX_NUM_COEFS];
-        size_t n = 0;
+        int n = 0;
         while(t >= _calTime) {
             try {
                 n = _calFile->readData(d,MAX_NUM_COEFS);
